requirements.txt
amqp
asgiref
attrs
autobahn
Automat
billiard
celery
certifi
cffi
channels
channels_redis
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
constantly
cryptography
cssselect2
daphne
Django
django-cors-headers
django-redis
djangorestframework
gunicorn
hyperlink
idna
incremental
kombu
lxml
modeltranslation
msgpack
packaging
pillow
prompt_toolkit
pyasn1
pyasn1_modules
pycparser
pyOpenSSL
python-dateutil
python-dotenv
redis
reportlab
requests
service-identity
setuptools
six
sqlparse
svglib
tinycss2
Twisted
txaio
typing_extensions
tzdata
urllib3
vine
wcwidth
webencodings
zope.interface
